The bash script and gsctl tests are complete
python file for gsctl

fix main.py
